requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
comm==0.2.2
debugpy==1.6.7
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.1.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
importlib_metadata==8.4.0
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jiter==0.5.0
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
multidict==6.0.5
nest_asyncio==1.6.0
numpy==2.1.0
openai==1.43.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pip==24.2
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
regex==2024.7.24
requests==2.32.3
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.2
tiktoken==0.7.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
wheel==0.43.0
yarl==1.9.7
zipp==3.20.1
